Working on day 5
I counted the inputs and decided that it would be faster to just goes through every map one by one. But the other half of the assignment makes it like billions of numbers.
It only requires a small code change, but just reading the numbers takes a lot of rime and resources and I implement sanity check.